rqe_iterators: move IdListLazy/MetricLazy into their own modules
Pure code move: extract the lazily-populated iterator variants out of
id_list.rs / metric.rs into dedicated id_list_lazy.rs / metric_lazy.rs
modules. No behavior change.
