+++ src/redisearch_rs/rqe_iterators/src/lib.rs
@@ -38,11 +38,13 @@ pub mod empty;
 pub mod expiration_checker;
 pub mod geo_shape;
 pub mod id_list;
+pub mod id_list_lazy;
 pub mod interop;
 pub mod intersection;
 pub mod inverted_index;
 pub mod maybe_empty;
 pub mod metric;
+pub mod metric_lazy;
 pub mod not;
 pub mod not_optimized;
 pub mod not_reducer;
